requirements.txt
Flask==3.1.0
Flask-WTF==1.2.1
WTForms==3.1.1
python-dotenv>=1.0.1
reportlab==4.1.0
requests>=2.32.0
langflow>=1.1.3
beautifulsoup4>=4.12.0
openai>=1.0.0
